+++ lib/language_operator/cli/commands/quickstart.rb
@@ -2,7 +2,7 @@
 
 require_relative '../base_command'
 require_relative '../formatters/progress_formatter'
-require_relative '../../ux/quickstart'
+require_relative '../wizards/quickstart_wizard'
 
 module LanguageOperator
   module CLI
@@ -11,7 +11,8 @@ module LanguageOperator
       class Quickstart < BaseCommand
         desc 'start', 'Interactive setup wizard for first-time users'
         def start
-          Ux::Quickstart.execute
+          wizard = Wizards::QuickstartWizard.new
+          wizard.run
         end
 
         default_task :start